+++ RELEASE_NOTES.md
@@ -1,3 +1,8 @@
+## v0.2.0 - 2021-06-xx
+
+- New widget: Tagger
+- `debounce` function modularized
+
 ## v0.1.0 - 2021-06-12
 
 - Initial release